// src/modules/angular/model/metric.ts
/**
 * Senso-Care
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SensorData } from './sensorData';


export interface Metric { 
    type?: string;
    sensors?: Array<SensorData>;
}